requirements.txt
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
decorator==4.4.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
instagrapi==2.2.1
moviepy==1.0.3
numpy==2.2.6
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.3.3
pillow==12.0.0
playwright==1.56.0
playwright-stealth==2.0.0
proglog==0.1.12
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.4
selenium==4.39.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
undetected-playwright==0.3.0
urllib3==2.6.0
websocket-client==1.9.0
wsproto==1.3.2
